pruebas unitarias del Get Select Contratar Mantenimientos #144 (#156)

+++ src/AppForSEII2526.API/DTOs/ReservaDTO/ReservaItemDTO.cs
@@ -3,12 +3,12 @@
     public class ReservaItemDTO
     {
         public int ReservaId { get; set; }
-        public decimal Price { get; set; }
+        public double Price { get; set; }
         public int NumberOfDays { get; set; }
         public string? Comentarios { get; set; }
         public string Name { get; set; }
 
-        public ReservaItemDTO(int reservaId, string name, decimal price, int numberofdays, string? comentarios= "")
+        public ReservaItemDTO(int reservaId, string name, double price, int numberofdays, string? comentarios= "")
         {
             ReservaId = reservaId;
             Name = name;
